broke settings page up into components
