+++ client/src/components/ShopManager/Settings/ShippingSettings.jsx
@@ -1,0 +1,110 @@
+import { useState } from 'react';
+
+const ShippingSettings = ({ address, setAddress, savedAddresses, setSavedAddresses, states, countries }) => {
+  const handleSaveAddress = () => {
+    setSavedAddresses([...savedAddresses, address]);
+    setAddress({ line1: '', line2: '', city: '', state: '', zip: '', country: 'USA' });
+  };
+
+  return (
+    <div>
+      <div className="mb-4">
+        <label className="block text-sm font-medium text-gray-700 mb-2">Address Line 1</label>
+        <input
+          type="text"
+          value={address.line1}
+          onChange={(e) => setAddress({ ...address, line1: e.target.value })}
+          className="block w-full mt-1 p-1 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+        />
+      </div>
+
+      <div className="mb-4">
+        <label className="block text-sm font-medium text-gray-700 mb-2">Address Line 2</label>
+        <input
+          type="text"
+          value={address.line2}
+          onChange={(e) => setAddress({ ...address, line2: e.target.value })}
+          className="block w-full mt-1 p-1 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+        />
+      </div>
+
+      <div className="mb-4">
+        <label className="block text-sm font-medium text-gray-700 mb-2">City</label>
+        <input
+          type="text"
+          value={address.city}
+          onChange={(e) => setAddress({ ...address, city: e.target.value })}
+          className="block w-full mt-1 p-1 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+        />
+      </div>
+
+      <div className="mb-4">
+        <label className="block text-sm font-medium text-gray-700 mb-2">State</label>
+        <select
+          value={address.state}
+          onChange={(e) => setAddress({ ...address, state: e.target.value })}
+          className="block w-full mt-1 p-1 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+        >
+          <option value="" disabled>
+            Select a state
+          </option>
+          {states.map((state, index) => (
+            <option key={index} value={state}>
+              {state}
+            </option>
+          ))}
+        </select>
+      </div>
+
+      <div className="mb-4">
+        <label className="block text-sm font-medium text-gray-700 mb-2">ZIP Code</label>
+        <input
+          type="text"
+          value={address.zip}
+          onChange={(e) => setAddress({ ...address, zip: e.target.value })}
+          className="block w-full mt-1 p-1 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+        />
+      </div>
+
+      <div className="mb-4">
+        <label className="block text-sm font-medium text-gray-700 mb-2">Country</label>
+        <select
+          value={address.country}
+          onChange={(e) => setAddress({ ...address, country: e.target.value })}
+          className="block w-full mt-1 p-1 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+        >
+          <option value="" disabled>
+            Select a country
+          </option>
+          {countries.map((country, index) => (
+            <option key={index} value={country}>
+              {country}
+            </option>
+          ))}
+        </select>
+      </div>
+
+      <div className="mb-4">
+        <button
+          onClick={handleSaveAddress}
+          className="px-4 py-2 rounded-md bg-black text-white text-sm hover:bg-gray-300 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2"
+        >
+          Save Address
+        </button>
+      </div>
+
+      <div className="mb-6">
+        <label className="block text-sm font-medium text-gray-700 mb-2">Saved Addresses</label>
+        <select className="block w-full mt-1 p-1 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500 sm:text-sm">
+          {savedAddresses.map((addr, index) => (
+            <option key={index} value={index}>
+              {`${addr.line1}, ${addr.line2} ${addr.city}, ${addr.state}, ${addr.country}`}
+            </option>
+          ))}
+        </select>
+      </div>
+    </div>
+  );
+};
+
+export default ShippingSettings;
